+++ src/main/kotlin/icu/windea/pls/localisation/codeInsight/completion/ParadoxCommandScopeCompletionProvider.kt
@@ -1,9 +1,11 @@
 package icu.windea.pls.localisation.codeInsight.completion
 
 import com.intellij.codeInsight.completion.*
+import com.intellij.psi.util.*
 import com.intellij.util.*
 import icu.windea.pls.*
-import icu.windea.pls.config.cwt.*
+import icu.windea.pls.config.cwt.CwtConfigHandler.completeLocalisationCommandScope
+import icu.windea.pls.localisation.psi.*
 
 /**
  * 提供命令字段作用域的代码补全。
@@ -11,12 +13,22 @@ import icu.windea.pls.config.cwt.*
 @Suppress("UnstableApiUsage")
 class ParadoxCommandScopeCompletionProvider : CompletionProvider<CompletionParameters>() {
 	override fun addCompletions(parameters: CompletionParameters, context: ProcessingContext, result: CompletionResultSet) {
+		val offsetInParent = parameters.offset - parameters.position.textRange.startOffset
+		val keyword = parameters.position.getKeyword(offsetInParent)
 		val file = parameters.originalFile
 		val project = file.project
-		
-		//提示scope
 		val gameType = file.fileInfo?.gameType ?: return
 		val configGroup = getCwtConfig(project).get(gameType) ?: return
-		CwtConfigHandler.completeLocalisationCommandScope(configGroup, result)
+		
+		context.put(PlsCompletionKeys.offsetInParentKey, offsetInParent)
+		context.put(PlsCompletionKeys.keywordKey, keyword)
+		
+		val prevScope = parameters.position.parent?.siblings(forward = false, withSelf = false)
+			?.find { it is ParadoxLocalisationCommandScope }
+			?.text
+		if(prevScope != null) context.put(PlsCompletionKeys.prevScopeKey, prevScope)
+		
+		//提示scope
+		context.completeLocalisationCommandScope(configGroup, result)
 	}
 }